feat: post modal/page for welcome post type
