+++ packages/webapp/components/layouts/ProfileLayout/index.tsx
@@ -166,7 +166,7 @@ export default function ProfileLayout({
           )}
         >
           <div className="flex tablet:flex-col items-center self-start tablet:px-2 tablet:pt-2 tablet:pb-4 mb-6 tablet:mb-0 rounded-2xl bg-theme-bg-secondary">
-            <ProfilePicture user={profile} size="xxxlarge" />
+            <ProfilePicture user={profile} size="profile" />
             <div className="flex flex-col tablet:items-center mx-6 tablet:mx-0 tablet:mt-4 typo-footnote">
               <a
                 href={reputationGuide}
