+++ packages/webapp/pages/squads/[handle]/index.tsx
@@ -41,6 +41,7 @@ import {
   useTutorial,
   TutorialKey,
 } from '@dailydotdev/shared/src/hooks/useTutorial';
+import { PostType } from '@dailydotdev/shared/src/graphql/posts';
 import { mainFeedLayoutProps } from '../../../components/layouts/MainFeedPage';
 import { getLayout } from '../../../components/layouts/FeedLayout';
 import ProtectedPage from '../../../components/ProtectedPage';
@@ -121,6 +122,7 @@ const SquadPage = ({ handle }: SourcePageProps): ReactElement => {
     () => ({
       source: squadId,
       ranking: 'TIME',
+      supportedTypes: [PostType.Article, PostType.Share, PostType.Welcome],
     }),
     [squadId],
   );
